backend/requirements.txt
flask
flask-cors
pyjwt
pillow
numpy
opencv-python-headless
reportlab
python-dotenv
requests
groq
supabase
bcrypt
gunicorn
